app2 _receiver/requirements.txt
pika==1.3.0 
pyyaml==5.4.1
redis==4.3.4
Flask==2.1.3